requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-spectacular==0.27.0
pydantic==2.5.0

# Voice Notes - Transcription (OpenAI API)
openai==1.12.0

# Voice Notes - Local Transcription (optional, for VOICE_NOTE_PROVIDER=local)
faster-whisper==1.0.1

# Voice Notes - Async Processing
celery==5.3.6
redis==5.0.1

# Voice Notes - Audio Duration (optional, ffprobe preferred)
mutagen==1.47.0